requirements-dev.txt
annotated-types
anyio
build
certifi
charset-normalizer
click
contourpy
coverage
cycler
exceptiongroup
fastapi
flake8
fonttools
h11
idna
importlib_metadata
importlib_resources
iniconfig
joblib
kiwisolver
matplotlib
mccabe
mypy
mypy_extensions
numpy
packaging
pillow
pluggy
pycodestyle
pydantic
pydantic_core
pyDOE
pyflakes
pyparsing
pyproject_hooks
pytest
pytest-cov
pytest-env
python-dateutil
PyYAML
requests
scikit-learn
scipy
setproctitle
six
sniffio
starlette
tabulate
threadpoolctl
tomli
tqdm
types-PyYAML
types-tabulate
typing-inspection
typing_extensions
urllib3
uvicorn
wrapt
zipp
